webexBot/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
django==4.2.3 ; python_version >= '3.8'
django-filter==23.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.7.1 ; python_version < '3.11'
tzdata==2023.3 ; sys_platform == 'win32'
